fix: using import() to load the fs
