+++ test/config/config.test.ts
@@ -1,6 +1,7 @@
 import { Config } from '../../src';
+import { readFileSync } from 'fs';
 
-const config = Config.newConfig('test/config/testini.ini');
+const config = Config.newConfigFromText(readFileSync('test/config/testini.ini').toString());
 
 describe('multi-line test', () => {
   it('should config.get("multi1::name") to equal r.sub==p.sub&&r.obj==p.obj', function () {
